move test aggregation circuit to dev
